# Conflicts:
#	src/main/java/ru/coffeecoders/questbot/services/GameService.java
#	src/test/java/ru/coffeecoders/questbot/services/GameServiceTest.java
#	src/test/java/ru/coffeecoders/questbot/services/QuestionGroupServiceTest.java

+++ src/test/java/ru/coffeecoders/questbot/services/TaskServiceTest.java
@@ -28,13 +28,13 @@ class TaskServiceTest {
     private TaskService service;
 
     private Task task;
-    private long id;
+    private int id;
     private String name;
 
     @BeforeEach
     void setUp() {
         task = new Task();
-        id = 111L;
+        id = 111;
         name = "Test";
     }
